commit
Deployment working now, needed to log in and create a docker secret
key. Next to tell minikube this is local development and to run the
command eval(minikube docker-env) at the start of the shell script.
This is required by shell instance. I have put a sleep in the bash
script to check when the deployment is running, and the other
terminating